+++ src/data/usecases/load-account-by-token/db-load-account-by-token.ts
@@ -1,7 +1,4 @@
-import { Decrypter } from '@/data/protocols/criptography/decrypter'
-import { LoadAccountByTokenRepository } from '@/data/protocols/db/account/load-account-by-token-repository'
-import { AccountModel } from '@/data/usecases/add-account/db-add-account-protocols'
-import { LoadAccountByToken } from '@/domain/usecases/load-account-by-token'
+import { AccountModel, Decrypter, LoadAccountByTokenRepository, LoadAccountByToken } from './db-load-account-by-token-protocols'
 
 export class DbLoadAccountByToken implements LoadAccountByToken {
   constructor (
