+++ src/data/usecases/load-survey-by-id/db-load-survey-by-id.spec.ts
@@ -1,7 +1,6 @@
 
 import { DbLoadSurveyById } from './db-load-survey-by-id'
-import { LoadSurveyByIdRepository } from '@/data/protocols/db/survey/load-survey-by-id-repository'
-import { SurveyModel } from '@/domain/models/survey'
+import { LoadSurveyByIdRepository, SurveyModel } from './db-load-survey-by-id-protocols'
 
 import MockDate from 'mockdate'
